rename pyam-analysis to pyam
