+++ pyam/timeseries.py
